backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.6.0
supabase>=2.0,<3.0